refactors to avoid circular dependency
FFTs now appear to be working smoothly
